refactor: everything separated into directories neatly -> can begin refactor
