+++ src/benchmarks/reachability.rs
@@ -2,8 +2,8 @@ use biodivine_lib_bdd::Bdd;
 use std::fmt::Debug;
 
 use crate::{
+    prototype::symbolic_domain::SymbolicDomain,
     prototype::{count_states, find_start_of, log_percent, pick_state_bdd, SmartSystemUpdateFn},
-    symbolic_domain::SymbolicDomain,
 };
 
 pub fn reachability_benchmark<D: SymbolicDomain<u8> + Debug>(sbml_path: &str) {
