+++ src/prototype/update_fn_bdd.rs
@@ -4,7 +4,7 @@ use biodivine_lib_bdd::{Bdd, BddPartialValuation, BddValuation, BddVariableSet};
 
 // use crate::{Expression, SymbolicDomain, UpdateFn};
 
-use crate::symbolic_domain::SymbolicDomain;
+use crate::prototype::symbolic_domain::SymbolicDomain;
 
 use super::{expression::Proposition, Expression, UpdateFn};
